Fix encoding problems causing lines to be commented
Some Shift-JIS characters turn to `\` when read as ASCII. When this is
at the end of the line of a comment, the next line will be commented
out, as well, causing major build issues.

This happens despite the -enc SJIS compiler flag, and is a byproduct of
WINE probably not doing a great job at supporting the Shift-JIS
encoding.  If a solution is ever found, this commit should be reverted.