Define USERNAME variable
Originally, this source would build slightly different things depending
on which user was logged in. This is probably due to how the different
developers worked on different parts of the game.

For us, this serves no purpose other than provide unpleasant surprises,
so it's for the best to define it as empty for now, allowing whoever is
building it to explicitly provide a username.

Eventually, everything should be built regardless of username.