No more differences, also fix header code, add fixrom, and add sha1 checking.
There seems to be an unused symbol at 0x021D0D60, as the symbols after it in the baserom are shifted by 0x4. It is thought to be unused as searching through an asm dump of platinum yields no result for that address, and the symbol before it (signal_funcs) which is an array would have shown diffs in the code if the array size was increased. However, since the build system currently cannot build the Metrowerks libraries, just increase the size of signal_funcs by 4 in the object file (located at 0x3358c in MSL_C_NITRO_Ai_LE.a) as well as the section size by 4 (located at 0x33be0). This should not cause any issues since it's bss and nothing actually accesses 0x021D0D60.