Fix lazy DWC c++ match in overlay 66.
The DWC c++ libraries in overlay 66 have been re-included in the build and the files modified to match. libdwcenc.a doesn't match, though, so the disassembled asm has to be included.

Also, the "unused variable" in MSL_C_NITRO_Ai_LE.a turned out to be just a variable used only by the DWC C++ libraries being deadstripped from the build. So revert back to the original MSL_C_NITRO_Ai_LE.a binary.

Barring missed NONEQUIVALENT/NONMATCHING functions with hardcoded offsets, the ROM should be shiftable.